Revert "Merge branch 'master' into Calvin-Mongo"
This reverts commit a4f4ccb17ebefefee29b1c196acbb2c0b198a1b0, reversing
changes made to 23acc9fc71d5432a9ac32fee77f1a626f340c0c2.

+++ src/main/java/base/studentFlowchartDb/Flowchart.java
@@ -13,13 +13,13 @@ public class Flowchart implements Serializable {
 
     @Id
     private String id;
-    private Long studentId;
+    private int studentId;
     private String flowchartName;
     private List<PlannedCourse> plannedCourses = new ArrayList<PlannedCourse>();
 
     public Flowchart() {}
 
-    public Flowchart(Long studentId, String name, List<PlannedCourse> plannedCourses) {
+    public Flowchart(int studentId, String name, List<PlannedCourse> plannedCourses) {
         this.studentId = studentId;
         this.flowchartName = name;
         this.plannedCourses.addAll(plannedCourses);
@@ -33,11 +33,11 @@ public class Flowchart implements Serializable {
         this.id = id;
     }
 
-    public Long getStudentId() {
+    public int getStudentId() {
         return studentId;
     }
 
-    public void setStudentId(Long studentId) {
+    public void setStudentId(int studentId) {
         this.studentId = studentId;
     }
